_make: Ensure children exit correctly on CTRL+C
unbuffer sucks. I'm not sure how exactly it works but it does not
propagate signals for some reason. So make wasn't getting SIGINT when
you'd press CTRL+C.

Not only would it not propagate the signal, it wouldn't wait for make
to exit and so as soon as you'd press CTRL+C your logs would be clipped
at that point. Now if you use a ./make.sh and press CTRL+C you'll see
error logs as things shut down.

The reason for unbuffer was to disable pipe buffering on make's stdout.
This way make's logs would become immediately available as they came in
to tee and then the background process writing them to .make-log.

Otherwise they come in jerky and the experience feels almost laggy.

I tried alternatives like stdbuf to see if they propagated signals
but stdbuf didn't even disable buffering. It works differently to
unbuffer and I suspect it just doesn't work on macOS as it's a GNU
tool.

Eventually I found script(1) which works perfectly without any
buffering issues and propagates signals correctly! It appears that any
redirection > or pipe | in a shell buffers but for some reason script
does not when writing to the fifo.

I also added _make_cleanup to ensure the background sed (and now also
tee) exit correctly.